+++ packages/components-angular/tests/vrt/specs/styles/focus.vrt.ts
@@ -1,5 +1,5 @@
 import { expect, test } from '@playwright/test';
-import { forceFocusState } from '../../../../../components-js/tests/vrt/playwright/helpers';
+import { forceFocusState, forceFocusVisibleState } from '../../../../../components-js/tests/vrt/playwright/helpers';
 import { viewportWidthM } from '@porsche-design-system/shared/testing/playwright.vrt.config';
 
 const style = 'styles-focus';
@@ -9,7 +9,9 @@ test.describe(style, async () => {
     await page.goto(`/${style}`);
     await page.setViewportSize({ width: viewportWidthM, height: 600 });
     await forceFocusState(page, 'a');
+    await forceFocusVisibleState(page, 'a');
     await forceFocusState(page, 'button');
+    await forceFocusVisibleState(page, 'button');
     await expect(page.locator('#app')).toHaveScreenshot(`${style}-${viewportWidthM}.png`);
   });
 });