+++ packages/components-js/tests/vrt/playwright/specs/states/textarea-wrapper.vrt.ts
@@ -3,6 +3,7 @@ import { schemes, themes, viewportWidthM } from '@porsche-design-system/shared/t
 import {
   forceFocusHoverState,
   forceFocusState,
+  forceFocusVisibleState,
   forceHoverState,
   getPlaygroundPseudoStatesMarkup,
   type PrefersColorScheme,
@@ -117,7 +118,7 @@ const scenario = async (page: Page, theme: Theme, scheme?: PrefersColorScheme): 
   await forceHoverState(page, '.hover p-textarea-wrapper.force-label >>> label');
   await forceHoverState(page, '.hover p-textarea-wrapper a');
   await forceFocusState(page, '.focus p-textarea-wrapper textarea');
-  await forceFocusState(page, '.focus p-textarea-wrapper a');
+  await forceFocusVisibleState(page, '.focus p-textarea-wrapper a');
   await forceFocusHoverState(page, '.focus-hover p-textarea-wrapper:not(.force-label) textarea');
   await forceFocusState(page, '.focus-hover p-textarea-wrapper.force-label textarea');
   await forceHoverState(page, '.focus-hover p-textarea-wrapper.force-label >>> label');