+++ packages/components/CHANGELOG.md
@@ -14,7 +14,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
+#### Added
+
 - `Select` ([#3008](https://github.com/porsche-design-system/porsche-design-system/pull/3008))
+
+#### Fixed
+
+- `Carousel`: Skip link is visible when it receives keyboard focus
+  ([#3055](https://github.com/porsche-design-system/porsche-design-system/pull/3055))
 
 ### [3.13.0-rc.0] - 2024-02-19